Move callbacks and main function into securedrop_proxy package
This is done for testability: the code in sd-proxy.py now just
uses the main function and callbacks which are in the package so
they can be unit tested.
